backend/requirements.txt
# FastAPI core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Redis
redis[hiredis]==5.2.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP client (for LLM API calls)
httpx==0.28.1

# Vector DB
qdrant-client==1.12.1

# Object storage
minio==7.2.12

# Validation & serialization
pydantic==2.10.4
pydantic-settings==2.7.1

# SSE streaming
sse-starlette==2.2.1
